Move status json cleaer into internal package
